refactor: profile module refactor
